+++ config.php
@@ -18,10 +18,39 @@ $all_links = $core->blog->settings->externallinks->all_links;
 $checkbox_new_links = $core->blog->settings->externallinks->checkbox_new_links;
 $one_link = $core->blog->settings->externallinks->one_link;
 $with_icon = $core->blog->settings->externallinks->with_icon;
+$new_icon = $core->blog->settings->externallinks->new_icon;
+$new_icon_file = null;
 
+$restore_url = $p_url . '&amp;restore_defaut_icon=1';
 $default_tab = 'externallinks_settings';
 
-if (!empty($_POST['saveconfig'])) {
+if (!empty($_SESSION['message'])) {
+    $message = $_SESSION['message'];
+    unset($_SESSION['message']);
+}
+
+if (!empty($_GET['restore_defaut_icon']) && $new_icon) {
+    $core->blog->settings->externallinks->put('new_icon', '', 'string');
+    $_SESSION['message'] = __('Default icon has been successfully restored');
+    http::redirect($p_url);
+} elseif (!empty($_POST['saveconfig'])) {
+    if (!empty($_FILES['new_icon']['tmp_name'])) {
+        try {
+            files::uploadStatus($_FILES['new_icon']);
+            $new_icon = strtolower(files::tidyFileName($_FILES['new_icon']['name']));
+            $core->media = new dcMedia($core);
+            $media_id = $core->media->uploadFile(
+                $_FILES['new_icon']['tmp_name'],
+                $new_icon
+            );
+            $core->blog->settings->externallinks->put('new_icon', $media_id, 'string');
+            $_SESSION['message'] = __('Icon successfully changed');
+            http::redirect($p_url);
+        } catch (Exception $e) {
+            $core->error->add($e->getMessage());
+        }
+    }
+
     try {
         $active = (empty($_POST['active']))?false:true;
         $all_links = (empty($_POST['all_links']))?false:true;
@@ -40,5 +69,11 @@ if (!empty($_POST['saveconfig'])) {
     } catch (Exception $e) {
         $core->error->add($e->getMessage());
     }
+} else {
+    if ($new_icon) {
+        $core->media = new dcMedia($core);
+        $new_icon_file = $core->media->getFile($new_icon)->file_url;
+    }
 }
+
 include(dirname(__FILE__) . '/tpl/index.tpl');